+++ package.json
@@ -11,6 +11,8 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^8.0.4",
+    "react-router": "^6.4.1",
+    "react-router-dom": "^6.4.1",
     "react-scripts": "4.0.3",
     "redux-saga": "^1.2.1",
     "web-vitals": "^1.1.2"
