+++ src/container/Home/Home.jsx
@@ -1,51 +1,11 @@
-import React, { useState, useEffect } from "react";
+import PropTypes from "prop-types";
 
-import Navbar from "components/Navbar/Navbar";
+function Home({ hideMobileProducts, isDesktop }) {
 
-function Home() {
-  const [hideMobileProducts, sethideMobileProducts] = useState(true);
-  const [isDesktop, setIsDesktop] = useState(false)
-
-  useEffect(() => {
-    window.addEventListener("resize", handleResize);
-  }, [])
-  /** 
-  checks if screen is desktop width, 
-  this is needed to render the desktopNavItems in the <NavBar /> Component
-   */
-  useEffect(() => {
-    if (window.innerWidth >= 1366) {
-      setIsDesktop(true)
-    }
-  }, [])
-
-  //resets mobile navigation state  
-  useEffect(() => {
-    sethideMobileProducts(true);
-  }, [isDesktop])
-
-  const handleResize = () => {
-
-    if (window.innerWidth >= 1366) {
-      setIsDesktop(true)
-    } else {
-      setIsDesktop(false);
-    }
-  }
-
-  const hideMobileProductsMenu = () => {
-
-    if (hideMobileProducts) {
-      sethideMobileProducts(false);
-    } else {
-      sethideMobileProducts(true);
-    }
-
-  };
 
   return (
     <div className="Home-container" data-testid="home-container">
-      <Navbar hideMobileProductsMenu={hideMobileProductsMenu} hideMobileProducts={hideMobileProducts} isDesktop={isDesktop} />
+      <p>Home</p>
       {hideMobileProducts || isDesktop ? null : (
         <div className="overlay">
         </div>
@@ -54,6 +14,9 @@ function Home() {
   );
 }
 
-Home.propTypes = {};
+Home.propTypes = {
+  hideMobileProducts: PropTypes.func,
+  isDesktop: PropTypes.bool
+};
 
 export default Home;
